+++ lib/multiple_table_inheritance.rb
@@ -3,5 +3,7 @@ require 'multiple_table_inheritance/railtie'
 module MultipleTableInheritance
   extend ActiveSupport::Autoload
   
-  autoload :ActiveRecord
+  autoload :Child
+  autoload :Parent
+  autoload :Migration
 end