+++ lib/multiple_table_inheritance/active_record.rb
@@ -1,9 +1,0 @@
-module MultipleTableInheritance
-  module ActiveRecord
-    extend ActiveSupport::Autoload
-    
-    autoload :Child
-    autoload :Parent
-    autoload :Migration
-  end
-end