+++ lib/multiple_table_inheritance/child.rb
@@ -1,0 +1,7 @@
+module MultipleTableInheritance
+  module Child
+    extend ActiveSupport::Autoload
+    
+    autoload :Base
+  end
+end