+++ lib/multiple_table_inheritance/parent.rb
@@ -1,0 +1,8 @@
+module MultipleTableInheritance
+  module Parent
+    extend ActiveSupport::Autoload
+    
+    autoload :Base
+    autoload :Relation
+  end
+end