+++ spec/spec_helper.rb
@@ -2,38 +2,33 @@ require 'active_record'
 require 'database_cleaner'
 require 'multiple_table_inheritance'
 
-MultipleTableInheritance::ActiveRecord::Parent::ClassMethods.module_eval do
-  def find_without_inheritance(&block)
-    # remove_method :find_by_sql
-    extend ActiveRecord::Querying
-    result = yield
-    extend MultipleTableInheritance::ActiveRecord::Parent::FinderMethods
-    result
-  end
-end
-
 MultipleTableInheritance::Railtie.insert
 
 require 'support/tables'
 require 'support/models'
 
 module MultipleTableInheritanceSpecHelper
-  def mock_everything
-    team = Team.create!(:name => 'Website')
-    java = Language.create!(:name => 'Java')
-    cpp = Language.create!(:name => 'C++')
-    Programmer.create!(
-        :first_name => 'Mario',
-        :last_name => 'Mario',
-        :salary => 65000,
-        :team => team,
-        :languages => [java, cpp])  # programmer-specific relationship
-    Manager.create!(
-        :first_name => 'King',
-        :last_name => 'Koopa',
-        :salary => 70000,
-        :team => team,
-        :bonus => 5000)  # manager-specific field
+  def mock_everything!
+    3.times do |i|
+      team = Team.create!(:name => "Team#{i}")
+      language = Language.create!(:name => "Java 1.#{i + 4}")
+    end
+    
+    3.times do |i|
+      Programmer.create!(
+          :first_name => "Joe",
+          :last_name => "Schmoe #{i}",
+          :salary => 60000 + (i * 5000),
+          :team => Team.first,
+          :languages => Language.limit(2))  # programmer-specific relationship
+      
+      Manager.create!(
+          :first_name => "Bob",
+          :last_name => "Smith #{i}",
+          :salary => 70000 + (i * 2500),
+          :team => Team.first,
+          :bonus => i * 2500)  # manager-specific field
+    end
   end
 end
 
@@ -52,7 +47,6 @@ RSpec.configure do |config|
   
   config.before(:each) do
     DatabaseCleaner.start
-    mock_everything
   end
   
   config.after(:each) do