+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+0.1.2
+
+* Fixed a bug that resulted in parent record not being accessible after
+  performing a find directly on the child model.
+* Added many more unit tests.
+
 0.1.1
 
 * Fixed bidirectional deletion of records.